requirements.txt
setuptools-rust
torch 
torchvision 
torchaudio
llama-cpp-python
SpeechRecognition[faster-whisper]
gpt4all
pyaudio
soundfile==0.12.1
pyttsx3
pyautogui
roboflow
opencv-python
ultralytics
python-doctr
ollama
